Update script.js
Updated player logic for drawing aces. Every time an ace is valued at 11 it changes a Boolean variable "aceEquals11" to true. Each draw now runs a function that checks if "aceEquals11" is true for any of the players cards and reduces the value of the hand by 10 for each instance.